+++ FAA/FAR25/__init__.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+Data = namedtuple('Data', 'W flag')