+++ CHANGELOG.md
@@ -4,8 +4,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [2.0.0] - 14-05-2018
-- FSBUILD import fsbroute.log file
-- Multiple arr/dep analysis
+### Added
+- FSBUILD import fsbroute.log route file
+- Arrival / departure airport analysis
 
 ## [1.0.1] - 16-03-2018
 ### Added